[frontend] move Semi-phase type machinery under semi::
Today's `Ty` is specifically the *Semi* type (it carries generics and
holes); the monomorphized *Full* type is a genuinely different
representation. The Semi/Full split is already known from the original
(`Data.Semi.Type` vs `Data.Full.Type`), so there is no reason to wait —
scope the Semi-phase machinery under `semi::` now, while it is all in one
PR, rather than leave a TODO.

Moves `ty`, `infer`, and `traits` (all Semi-phase: the Semi type, its
unification solver, and the trait system that resolves over it) under
`crate::semi`, alongside the elaboration passes. `surface` (parser-facing
input) and `utils` stay top-level. `full::*` will be its sibling later.

Pure restructuring: git-tracked renames plus import-path rewrites; no
behavior change. 31 tests pass, clippy clean.
